switch to ipfs:// URLs
Apparently some online sources say that opensea requires ipfs:// URLs
rather than proxies but i'm skeptical since our erc721 contract didn't
do that and it worked.  opensea had some issues loading the nft metadata
so tried to change this to see if it worked.  either way, looks like
opensea is fixed now and things work so keeping it as is.